+++ src/lib.rs
@@ -2,7 +2,6 @@ use app::App;
 
 mod api;
 mod app;
-mod auth;
 mod components;
 mod database_queries;
 mod errors;
